10.29 Initialize

+++ src/main/java/com/zht/newclassmanager/context/BaseContext.java
@@ -1,13 +1,13 @@
 package com.zht.newclassmanager.context;
 
 public class BaseContext {
-    public static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    public static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
